fix media old_path

+++ src/Http/Controllers/Controller.php
@@ -170,8 +170,11 @@ abstract class Controller extends BaseController
             $data->save();
             if ($old_path != $new_path && !Storage::disk(config('voyager.storage.disk'))->exists($new_path)) {
                 $request->session()->forget([$slug.'_path', $slug.'_uuid']);
-                Storage::disk(config('voyager.storage.disk'))->move($old_path, $new_path);
-                Storage::disk(config('voyager.storage.disk'))->deleteDirectory($folder_path);
+                if(Storage::disk(config('voyager.storage.disk'))->exists($old_path))
+                {
+                    Storage::disk(config('voyager.storage.disk'))->move($old_path, $new_path);
+                    Storage::disk(config('voyager.storage.disk'))->deleteDirectory($folder_path);
+                }
             }
         }
